Isolates price info since its based in config values
